Made restserver work properly
The rest server is now working, so you can see the given data in localhost8080.
Made a RecipeHubModel class, which was needed for the REST API.
Applied necessary changes to Controller and Service to use RecipeHubModel.
Made necessary changes for Gson to work.
Password-hashing now happens when making a new Profile object.
Since writeFile() was a made static, writeRecipeLibrary() in RecipeFilehandler needed small change.

Issue: #82